Fix contextual menu's button style.
The buttons in the context menu that is being displayed when
selecting text had a white color. As the menu's background is
white, as well, the buttons were not visible at all.

It turns out this is a bug in the Holo.Light.DarkActionBar theme
(see https://code.google.com/p/android/issues/detail?id=26008).
Change parent theme to Holo.Light to fix the issue.